bot/requirements.txt
nonebot2~=2.0.0a14
nonebot-adapter-cqhttp~=2.0.0a14
pydantic~=1.8.2
PyYAML~=5.4.1
ruamel.yaml~=0.17.16
lxml~=4.6.3
httpx~=0.18.2
pymongo~=3.12.0
bilibili-api~=8.3.1
peewee==3.14.8
pymysql==1.0.2
APScheduler~=3.8.1